+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 2.13
 ----
 
-* Not yet released.
+* Released on 17th May 2022.
 * Dropped support for Python 3.6.
 * Improved detection of PWG 5100.13 based strings files.
 * Added support for ResourceDictionary format.